+++ EmployeeAssets/src/main/java/com/project3/service/EmployeeService.java
@@ -6,9 +6,9 @@ import com.project3.model.Employee;
 
 public interface EmployeeService {
 	
-	void deleteEmployee(Long id);
+	void deleteEmployee(int id);
 	List<Employee> findAllEmployees();
-    Employee findEmployeeById(Long id);
+    Employee findEmployeeById(int id);
     Employee save(Employee employee);
 	void updateEmployee(Employee employee);
 }
